packages/forge-media/src/forge_media/api/routes.py
"""FastAPI routes for media processing endpoints."""

from __future__ import annotations

from typing import Literal

from fastapi import APIRouter, File, HTTPException, UploadFile
from pydantic import BaseModel

from forge_media.core.media_set import MediaSet, MediaItem, ProcessingConfig
from forge_media.documents.extractors.base import ExtractedDocument, ExtractedTable
from forge_media.documents.chunking import TextChunk
from forge_media.vision.ocr import OCRResult

router = APIRouter(prefix="/api/v1/media", tags=["media"])


# ============================================================================
# Request/Response Models
# ============================================================================


class CreateMediaSetRequest(BaseModel):
    """Request to create a new media set."""

    name: str
    description: str | None = None
    config: ProcessingConfig | None = None


class ExtractDocumentResponse(BaseModel):
    """Response from document extraction."""

    full_text: str
    page_count: int
    word_count: int
    tables: list[ExtractedTable]
    metadata: dict


class ChunkDocumentResponse(BaseModel):
    """Response from document chunking."""

    chunks: list[TextChunk]
    total_chunks: int


class OCRResponse(BaseModel):
    """Response from OCR processing."""

    full_text: str
    confidence: float
    language: str | None


# ============================================================================
# Media Set Endpoints
# ============================================================================


@router.post("/sets", response_model=MediaSet)
async def create_media_set(request: CreateMediaSetRequest) -> MediaSet:
    """
    Create a new media set.

    Media sets are collections of related media items with shared
    processing configuration.
    """
    # Stub implementation - would interact with database
    raise HTTPException(status_code=501, detail="Not implemented")


@router.get("/sets", response_model=list[MediaSet])
async def list_media_sets(
    limit: int = 50,
    offset: int = 0,
) -> list[MediaSet]:
    """List media sets for the current user."""
    raise HTTPException(status_code=501, detail="Not implemented")


@router.get("/sets/{set_id}", response_model=MediaSet)
async def get_media_set(set_id: str) -> MediaSet:
    """Get a specific media set by ID."""
    raise HTTPException(status_code=501, detail="Not implemented")


@router.delete("/sets/{set_id}")
async def delete_media_set(set_id: str) -> dict:
    """Delete a media set and all its items."""
    raise HTTPException(status_code=501, detail="Not implemented")


# ============================================================================
# Media Item Endpoints
# ============================================================================


@router.post("/sets/{set_id}/items", response_model=MediaItem)
async def upload_media(
    set_id: str,
    file: UploadFile = File(...),
    process: bool = True,
) -> MediaItem:
    """
    Upload a media item to a set.

    Args:
        set_id: Target media set ID
        file: File to upload
        process: Whether to process immediately
    """
    raise HTTPException(status_code=501, detail="Not implemented")


@router.get("/sets/{set_id}/items", response_model=list[MediaItem])
async def list_media_items(
    set_id: str,
    limit: int = 50,
    offset: int = 0,
) -> list[MediaItem]:
    """List items in a media set."""
    raise HTTPException(status_code=501, detail="Not implemented")


@router.get("/items/{item_id}", response_model=MediaItem)
async def get_media_item(item_id: str) -> MediaItem:
    """Get a specific media item."""
    raise HTTPException(status_code=501, detail="Not implemented")


@router.delete("/items/{item_id}")
async def delete_media_item(item_id: str) -> dict:
    """Delete a media item."""
    raise HTTPException(status_code=501, detail="Not implemented")


@router.post("/items/{item_id}/reprocess", response_model=MediaItem)
async def reprocess_media(
    item_id: str,
    config: ProcessingConfig | None = None,
) -> MediaItem:
    """Reprocess a media item with new configuration."""
    raise HTTPException(status_code=501, detail="Not implemented")


# ============================================================================
# Document Extraction Endpoints
# ============================================================================


@router.post("/documents/extract", response_model=ExtractDocumentResponse)
async def extract_document(
    file: UploadFile = File(...),
    extract_tables: bool = True,
    extract_images: bool = False,
) -> ExtractDocumentResponse:
    """
    Extract content from a document.

    Supports PDF, DOCX, and XLSX files.
    """
    raise HTTPException(status_code=501, detail="Not implemented")


@router.post("/documents/chunk", response_model=ChunkDocumentResponse)
async def chunk_document(
    file: UploadFile = File(...),
    strategy: Literal["fixed", "semantic", "page"] = "semantic",
    chunk_size: int = 1000,
    overlap: int = 200,
) -> ChunkDocumentResponse:
    """
    Extract and chunk a document.

    Chunks are suitable for embedding and RAG pipelines.
    """
    raise HTTPException(status_code=501, detail="Not implemented")


@router.post("/documents/tables", response_model=list[ExtractedTable])
async def extract_tables(
    file: UploadFile = File(...),
) -> list[ExtractedTable]:
    """Extract tables from a document."""
    raise HTTPException(status_code=501, detail="Not implemented")


# ============================================================================
# Vision Endpoints
# ============================================================================


@router.post("/vision/ocr", response_model=OCRResponse)
async def run_ocr(
    file: UploadFile = File(...),
    languages: list[str] | None = None,
    provider: Literal["tesseract", "cloud_vision"] = "tesseract",
) -> OCRResponse:
    """
    Run OCR on an image.

    Args:
        file: Image file
        languages: Language codes (e.g., ["en", "es"])
        provider: OCR provider to use
    """
    raise HTTPException(status_code=501, detail="Not implemented")


@router.post("/vision/thumbnail")
async def generate_thumbnail(
    file: UploadFile = File(...),
    width: int = 256,
    height: int = 256,
) -> bytes:
    """Generate a thumbnail from an image."""
    raise HTTPException(status_code=501, detail="Not implemented")
