packages/forge-media/pyproject.toml
[project]
name = "open-forge-media"
version = "0.1.0"
description = "Media processing for Open Forge: document extraction, OCR, and image processing"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0.0",
    # Document extraction
    "pypdf>=4.0.0",
    "python-docx>=1.1.0",
    "openpyxl>=3.1.0",
    # Image processing
    "pillow>=10.0.0",
    # OCR
    "pytesseract>=0.3.10",
    # MIME type detection
    "python-magic>=0.4.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.8.0",
    "ruff>=0.2.0",
]
# Cloud vision providers (optional)
cloud = [
    "google-cloud-vision>=3.5.0",
    "boto3>=1.34.0",  # AWS Textract/Rekognition
]
# Advanced document processing
advanced = [
    "pdfplumber>=0.10.0",
    "pymupdf>=1.23.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/forge_media"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
strict = true
plugins = ["pydantic.mypy"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
